requirements.txt
absl-py==2.3.1
axial_positional_embedding==0.3.12
boto3==1.39.13
botocore==1.39.13
click==8.2.1
cloudpickle==3.1.1
CoLT5-attention==0.11.1
cramjam==2.10.0
dask==2025.7.0
einops==0.8.1
fastparquet==2024.11.0
fsspec==2024.6.1
geopandas==1.1.1
grpcio==1.73.1
hyper-connections==0.2.1
jmespath==1.0.1
local-attention==1.11.2
matplotlib==3.10.0
mkl-service==2.4.0
nbsphinx==0.9.7
nbsphinx-link==1.3.1
partd==1.4.2
product_key_memory==0.2.11
protobuf==6.31.1
pyarrow==20.0.0
pyogrio==0.11.0
pyproj==3.7.1
PyQt6==6.7.1
reformer-pytorch==1.4.4
s3transfer==0.13.1
seaborn==0.13.2
shapely==2.1.1
sphinx-rtd-theme==3.0.2
sphinxcontrib-jquery==4.1
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
toolz==1.0.0
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
triton==3.3.1
webencodings==0.5.1
Werkzeug==3.1.3
xlsxwriter==3.2.9
